// packages/server/src/generated/types.gen.ts
export type Webhooks =
  | OnComponentCreatedWebhookRequest
  | OnComponentDeletedWebhookRequest
  | OnComponentUpdatedWebhookRequest
  | OnCustomerCreatedWebhookRequest
  | OnCustomerCreationRejectedWebhookRequest
  | OnCustomerDeletedWebhookRequest
  | OnCustomerUpdatedWebhookRequest
  | OnDeploymentZoneCreatedWebhookRequest
  | OnDeploymentZoneDeletedWebhookRequest
  | OnReleaseDeployedWebhookRequest
  | OnDeploymentZoneUpdatedWebhookRequest
  | OnEntitlementCreatedWebhookRequest
  | OnEntitlementDeletedWebhookRequest
  | OnEntitlementUpdatedWebhookRequest
  | OnEntitlementGroupCreatedWebhookRequest
  | OnEntitlementGroupDeletedWebhookRequest
  | OnEntitlementGroupUpdatedWebhookRequest
  | OnFeatureFlagCreatedWebhookRequest
  | OnFeatureFlagDeletedWebhookRequest
  | OnFeatureFlagEvaluatedWebhookRequest
  | OnFeatureFlagUpdatedWebhookRequest
  | OnSystemOrganizationTokenIssuedWebhookRequest
  | OnInstanceEntitlementCapExceededWebhookRequest
  | OnInstanceEntitlementUsagePeriodRolledOverWebhookRequest
  | OnInstanceEntitlementUsageReachedWebhookRequest
  | OnInstanceEntitlementUsageReportAcceptedWebhookRequest
  | OnInstanceEntitlementUsageReportRejectedWebhookRequest
  | OnInstanceEntitlementUsageWarningThresholdReachedWebhookRequest
  | OnInstanceEntitlementValueGetWebhookRequest
  | OnInstanceCreatedWebhookRequest
  | OnInstanceDeletedWebhookRequest
  | OnInstanceDeployedWebhookRequest
  | OnInstanceLifecycleStageChangedWebhookRequest
  | OnInstanceMigratedWebhookRequest
  | OnInstanceStatusChangedWebhookRequest
  | OnInstanceUpdatedWebhookRequest
  | OnLicenseEntitlementAssignedWebhookRequest
  | OnLicenseEntitlementUnassignedWebhookRequest
  | OnLicenseEntitlementUpdatedWebhookRequest
  | OnLicenseArchivedWebhookRequest
  | OnLicenseCreatedWebhookRequest
  | OnLicenseDeletedWebhookRequest
  | OnLicensePublishedWebhookRequest
  | OnLicenseUnarchivedWebhookRequest
  | OnLicenseUpdatedWebhookRequest
  | OnLicenseFamilyCreatedWebhookRequest
  | OnLicenseFamilyDeletedWebhookRequest
  | OnLicenseFamilyUpdatedWebhookRequest
  | OnMetadataFieldArchivedWebhookRequest
  | OnMetadataFieldCreatedWebhookRequest
  | OnMetadataFieldReorderedWebhookRequest
  | OnMetadataFieldUnarchivedWebhookRequest
  | OnMetadataFieldUpdatedWebhookRequest
  | OnReleaseCreatedWebhookRequest
  | OnReleaseDeletedWebhookRequest;
// ...
export type OnComponentCreatedWebhookRequest = {
  body: OnComponentCreatedWebhookPayload;
  key: "com.kaiten.component.v1.created";
  path?: never;
  query?: never;
};

export type OnComponentDeletedWebhookPayload = {
  data: ComponentWritable;
  /**
   * Event name
   */
  name: "COMPONENT_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.component.v1.deleted";
};

export type OnComponentDeletedWebhookRequest = {
  body: OnComponentDeletedWebhookPayload;
  key: "com.kaiten.component.v1.deleted";
  path?: never;
  query?: never;
};

export type OnComponentUpdatedWebhookPayload = {
  data: ComponentWritable;
  /**
   * Event name
   */
  name: "COMPONENT_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.component.v1.updated";
};

export type OnComponentUpdatedWebhookRequest = {
  body: OnComponentUpdatedWebhookPayload;
  key: "com.kaiten.component.v1.updated";
  path?: never;
  query?: never;
};

export type OnCustomerCreatedWebhookPayload = {
  data: CustomerWritable;
  /**
   * Event name
   */
  name: "CUSTOMER_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.customer.v1.created";
};

export type OnCustomerCreatedWebhookRequest = {
  body: OnCustomerCreatedWebhookPayload;
  key: "com.kaiten.customer.v1.created";
  path?: never;
  query?: never;
};

export type OnCustomerCreationRejectedWebhookPayload = {
  data: CustomerCreationRejected;
  /**
   * Event name
   */
  name: "CUSTOMER_CREATION_REJECTED";
  /**
   * Event type
   */
  type: "com.kaiten.customer.v1.creation_rejected";
};

export type OnCustomerCreationRejectedWebhookRequest = {
  body: OnCustomerCreationRejectedWebhookPayload;
  key: "com.kaiten.customer.v1.creation_rejected";
  path?: never;
  query?: never;
};

export type OnCustomerDeletedWebhookPayload = {
  data: CustomerWritable;
  /**
   * Event name
   */
  name: "CUSTOMER_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.customer.v1.deleted";
};

export type OnCustomerDeletedWebhookRequest = {
  body: OnCustomerDeletedWebhookPayload;
  key: "com.kaiten.customer.v1.deleted";
  path?: never;
  query?: never;
};

export type OnCustomerUpdatedWebhookPayload = {
  data: CustomerWritable;
  /**
   * Event name
   */
  name: "CUSTOMER_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.customer.v1.updated";
};

export type OnCustomerUpdatedWebhookRequest = {
  body: OnCustomerUpdatedWebhookPayload;
  key: "com.kaiten.customer.v1.updated";
  path?: never;
  query?: never;
};

export type OnDeploymentZoneCreatedWebhookPayload = {
  data: DeploymentZoneWritable;
  /**
   * Event name
   */
  name: "DEPLOYMENT_ZONE_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.deployment_zone.v1.created";
};

export type OnDeploymentZoneCreatedWebhookRequest = {
  body: OnDeploymentZoneCreatedWebhookPayload;
  key: "com.kaiten.deployment_zone.v1.created";
  path?: never;
  query?: never;
};

export type OnDeploymentZoneDeletedWebhookPayload = {
  data: DeploymentZoneWritable;
  /**
   * Event name
   */
  name: "DEPLOYMENT_ZONE_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.deployment_zone.v1.deleted";
};

export type OnDeploymentZoneDeletedWebhookRequest = {
  body: OnDeploymentZoneDeletedWebhookPayload;
  key: "com.kaiten.deployment_zone.v1.deleted";
  path?: never;
  query?: never;
};

export type OnReleaseDeployedWebhookPayload = {
  data: Deployment;
  /**
   * Event name
   */
  name: "RELEASE_DEPLOYED";
  /**
   * Event type
   */
  type: "com.kaiten.deployment_zone.v1.release_deployed";
};

export type OnReleaseDeployedWebhookRequest = {
  body: OnReleaseDeployedWebhookPayload;
  key: "com.kaiten.deployment_zone.v1.release_deployed";
  path?: never;
  query?: never;
};

export type OnDeploymentZoneUpdatedWebhookPayload = {
  data: DeploymentZoneWritable;
  /**
   * Event name
   */
  name: "DEPLOYMENT_ZONE_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.deployment_zone.v1.updated";
};

export type OnDeploymentZoneUpdatedWebhookRequest = {
  body: OnDeploymentZoneUpdatedWebhookPayload;
  key: "com.kaiten.deployment_zone.v1.updated";
  path?: never;
  query?: never;
};

export type OnEntitlementCreatedWebhookPayload = {
  data: EntitlementWritable;
  /**
   * Event name
   */
  name: "ENTITLEMENT_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.entitlement.v1.created";
};

export type OnEntitlementCreatedWebhookRequest = {
  body: OnEntitlementCreatedWebhookPayload;
  key: "com.kaiten.entitlement.v1.created";
  path?: never;
  query?: never;
};

export type OnEntitlementDeletedWebhookPayload = {
  data: EntitlementWritable;
  /**
   * Event name
   */
  name: "ENTITLEMENT_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.entitlement.v1.deleted";
};

export type OnEntitlementDeletedWebhookRequest = {
  body: OnEntitlementDeletedWebhookPayload;
  key: "com.kaiten.entitlement.v1.deleted";
  path?: never;
  query?: never;
};

export type OnEntitlementUpdatedWebhookPayload = {
  data: EntitlementWritable;
  /**
   * Event name
   */
  name: "ENTITLEMENT_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.entitlement.v1.updated";
};

export type OnEntitlementUpdatedWebhookRequest = {
  body: OnEntitlementUpdatedWebhookPayload;
  key: "com.kaiten.entitlement.v1.updated";
  path?: never;
  query?: never;
};

export type OnEntitlementGroupCreatedWebhookPayload = {
  data: EntitlementGroupWritable;
  /**
   * Event name
   */
  name: "ENTITLEMENT_GROUP_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.entitlement_group.v1.created";
};

export type OnEntitlementGroupCreatedWebhookRequest = {
  body: OnEntitlementGroupCreatedWebhookPayload;
  key: "com.kaiten.entitlement_group.v1.created";
  path?: never;
  query?: never;
};

export type OnEntitlementGroupDeletedWebhookPayload = {
  data: EntitlementGroupWritable;
  /**
   * Event name
   */
  name: "ENTITLEMENT_GROUP_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.entitlement_group.v1.deleted";
};

export type OnEntitlementGroupDeletedWebhookRequest = {
  body: OnEntitlementGroupDeletedWebhookPayload;
  key: "com.kaiten.entitlement_group.v1.deleted";
  path?: never;
  query?: never;
};

export type OnEntitlementGroupUpdatedWebhookPayload = {
  data: EntitlementGroupWritable;
  /**
   * Event name
   */
  name: "ENTITLEMENT_GROUP_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.entitlement_group.v1.updated";
};

export type OnEntitlementGroupUpdatedWebhookRequest = {
  body: OnEntitlementGroupUpdatedWebhookPayload;
  key: "com.kaiten.entitlement_group.v1.updated";
  path?: never;
  query?: never;
};

export type OnFeatureFlagCreatedWebhookPayload = {
  data: FeatureFlagWritable;
  /**
   * Event name
   */
  name: "FEATURE_FLAG_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.feature_flag.v1.created";
};

export type OnFeatureFlagCreatedWebhookRequest = {
  body: OnFeatureFlagCreatedWebhookPayload;
  key: "com.kaiten.feature_flag.v1.created";
  path?: never;
  query?: never;
};

export type OnFeatureFlagDeletedWebhookPayload = {
  data: FeatureFlagWritable;
  /**
   * Event name
   */
  name: "FEATURE_FLAG_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.feature_flag.v1.deleted";
};

export type OnFeatureFlagDeletedWebhookRequest = {
  body: OnFeatureFlagDeletedWebhookPayload;
  key: "com.kaiten.feature_flag.v1.deleted";
  path?: never;
  query?: never;
};

export type OnFeatureFlagEvaluatedWebhookPayload = {
  data: FeatureFlagEvaluated;
  /**
   * Event name
   */
  name: "FEATURE_FLAG_EVALUATED";
  /**
   * Event type
   */
  type: "com.kaiten.feature_flag.v1.evaluated";
};

export type OnFeatureFlagEvaluatedWebhookRequest = {
  body: OnFeatureFlagEvaluatedWebhookPayload;
  key: "com.kaiten.feature_flag.v1.evaluated";
  path?: never;
  query?: never;
};

export type OnFeatureFlagUpdatedWebhookPayload = {
  data: FeatureFlagWritable;
  /**
   * Event name
   */
  name: "FEATURE_FLAG_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.feature_flag.v1.updated";
};

export type OnFeatureFlagUpdatedWebhookRequest = {
  body: OnFeatureFlagUpdatedWebhookPayload;
  key: "com.kaiten.feature_flag.v1.updated";
  path?: never;
  query?: never;
};

export type OnSystemOrganizationTokenIssuedWebhookPayload = {
  data: SystemTokenIssuance;
  /**
   * Event name
   */
  name: "SYSTEM_ORGANIZATION_TOKEN_ISSUED";
  /**
   * Event type
   */
  type: "com.kaiten.identity.v1.system_token_issued";
};

export type OnSystemOrganizationTokenIssuedWebhookRequest = {
  body: OnSystemOrganizationTokenIssuedWebhookPayload;
  key: "com.kaiten.identity.v1.system_token_issued";
  path?: never;
  query?: never;
};

export type OnInstanceEntitlementCapExceededWebhookPayload = {
  data: InstanceEntitlementCapExceeded;
  /**
   * Event name
   */
  name: "INSTANCE_ENTITLEMENT_CAP_EXCEEDED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.entitlement.v1.cap_exceeded";
};

export type OnInstanceEntitlementCapExceededWebhookRequest = {
  body: OnInstanceEntitlementCapExceededWebhookPayload;
  key: "com.kaiten.instance.entitlement.v1.cap_exceeded";
  path?: never;
  query?: never;
};

export type OnInstanceEntitlementUsagePeriodRolledOverWebhookPayload = {
  data: InstanceEntitlementUsagePeriodRolledOver;
  /**
   * Event name
   */
  name: "INSTANCE_ENTITLEMENT_USAGE_PERIOD_ROLLED_OVER";
  /**
   * Event type
   */
  type: "com.kaiten.instance.entitlement.v1.usage_period_rolled_over";
};

export type OnInstanceEntitlementUsagePeriodRolledOverWebhookRequest = {
  body: OnInstanceEntitlementUsagePeriodRolledOverWebhookPayload;
  key: "com.kaiten.instance.entitlement.v1.usage_period_rolled_over";
  path?: never;
  query?: never;
};

export type OnInstanceEntitlementUsageReachedWebhookPayload = {
  data: EntitlementUsageWritable;
  /**
   * Event name
   */
  name: "INSTANCE_ENTITLEMENT_USAGE_REACHED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.entitlement.v1.usage_reached";
};

export type OnInstanceEntitlementUsageReachedWebhookRequest = {
  body: OnInstanceEntitlementUsageReachedWebhookPayload;
  key: "com.kaiten.instance.entitlement.v1.usage_reached";
  path?: never;
  query?: never;
};

export type OnInstanceEntitlementUsageReportAcceptedWebhookPayload = {
  data: InstanceEntitlementUsageReportAccepted;
  /**
   * Event name
   */
  name: "ENTITLEMENT_USAGE_REPORT_ACCEPTED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.entitlement.v1.usage_report_accepted";
};

export type OnInstanceEntitlementUsageReportAcceptedWebhookRequest = {
  body: OnInstanceEntitlementUsageReportAcceptedWebhookPayload;
  key: "com.kaiten.instance.entitlement.v1.usage_report_accepted";
  path?: never;
  query?: never;
};

export type OnInstanceEntitlementUsageReportRejectedWebhookPayload = {
  data: InstanceEntitlementUsageReportAccepted;
  /**
   * Event name
   */
  name: "ENTITLEMENT_USAGE_REPORT_REJECTED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.entitlement.v1.usage_report_rejected";
};

export type OnInstanceEntitlementUsageReportRejectedWebhookRequest = {
  body: OnInstanceEntitlementUsageReportRejectedWebhookPayload;
  key: "com.kaiten.instance.entitlement.v1.usage_report_rejected";
  path?: never;
  query?: never;
};

export type OnInstanceEntitlementUsageWarningThresholdReachedWebhookPayload = {
  data: InstanceEntitlementUsageWarningThresholdReached;
  /**
   * Event name
   */
  name: "INSTANCE_ENTITLEMENT_USAGE_WARNING_THRESHOLD_REACHED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.entitlement.v1.usage_warning_threshold_reached";
};

export type OnInstanceEntitlementUsageWarningThresholdReachedWebhookRequest = {
  body: OnInstanceEntitlementUsageWarningThresholdReachedWebhookPayload;
  key: "com.kaiten.instance.entitlement.v1.usage_warning_threshold_reached";
  path?: never;
  query?: never;
};

export type OnInstanceEntitlementValueGetWebhookPayload = {
  data: InstanceEntitlementValueGetWritable;
  /**
   * Event name
   */
  name: "ENTITLEMENT_VALUE_GET";
  /**
   * Event type
   */
  type: "com.kaiten.instance.entitlement.v1.value_get";
};

export type OnInstanceEntitlementValueGetWebhookRequest = {
  body: OnInstanceEntitlementValueGetWebhookPayload;
  key: "com.kaiten.instance.entitlement.v1.value_get";
  path?: never;
  query?: never;
};

export type OnInstanceCreatedWebhookPayload = {
  data: InstanceWritable;
  /**
   * Event name
   */
  name: "INSTANCE_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.v1.created";
};

export type OnInstanceCreatedWebhookRequest = {
  body: OnInstanceCreatedWebhookPayload;
  key: "com.kaiten.instance.v1.created";
  path?: never;
  query?: never;
};

export type OnInstanceDeletedWebhookPayload = {
  data: InstanceWritable;
  /**
   * Event name
   */
  name: "INSTANCE_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.v1.deleted";
};

export type OnInstanceDeletedWebhookRequest = {
  body: OnInstanceDeletedWebhookPayload;
  key: "com.kaiten.instance.v1.deleted";
  path?: never;
  query?: never;
};

export type OnInstanceDeployedWebhookPayload = {
  data: InstanceWritable;
  /**
   * Event name
   */
  name: "INSTANCE_DEPLOYED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.v1.deployed";
};

export type OnInstanceDeployedWebhookRequest = {
  body: OnInstanceDeployedWebhookPayload;
  key: "com.kaiten.instance.v1.deployed";
  path?: never;
  query?: never;
};

export type OnInstanceLifecycleStageChangedWebhookPayload = {
  data: InstanceLifecycleStageChanged;
  /**
   * Event name
   */
  name: "INSTANCE_LIFECYCLE_STAGE_CHANGED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.v1.lifecycle_stage_changed";
};

export type OnInstanceLifecycleStageChangedWebhookRequest = {
  body: OnInstanceLifecycleStageChangedWebhookPayload;
  key: "com.kaiten.instance.v1.lifecycle_stage_changed";
  path?: never;
  query?: never;
};

export type OnInstanceMigratedWebhookPayload = {
  data: InstanceWritable;
  /**
   * Event name
   */
  name: "INSTANCE_MIGRATED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.v1.migrated";
};

export type OnInstanceMigratedWebhookRequest = {
  body: OnInstanceMigratedWebhookPayload;
  key: "com.kaiten.instance.v1.migrated";
  path?: never;
  query?: never;
};

export type OnInstanceStatusChangedWebhookPayload = {
  data: InstanceStatusChanged;
  /**
   * Event name
   */
  name: "INSTANCE_STATUS_CHANGED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.v1.status_changed";
};

export type OnInstanceStatusChangedWebhookRequest = {
  body: OnInstanceStatusChangedWebhookPayload;
  key: "com.kaiten.instance.v1.status_changed";
  path?: never;
  query?: never;
};

export type OnInstanceUpdatedWebhookPayload = {
  data: InstanceWritable;
  /**
   * Event name
   */
  name: "INSTANCE_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.instance.v1.updated";
};

export type OnInstanceUpdatedWebhookRequest = {
  body: OnInstanceUpdatedWebhookPayload;
  key: "com.kaiten.instance.v1.updated";
  path?: never;
  query?: never;
};

export type OnLicenseEntitlementAssignedWebhookPayload = {
  data: LicenseEntitlementWritable;
  /**
   * Event name
   */
  name: "LICENSE_ENTITLEMENT_ASSIGNED";
  /**
   * Event type
   */
  type: "com.kaiten.license.entitlement.v1.assigned";
};

export type OnLicenseEntitlementAssignedWebhookRequest = {
  body: OnLicenseEntitlementAssignedWebhookPayload;
  key: "com.kaiten.license.entitlement.v1.assigned";
  path?: never;
  query?: never;
};

export type OnLicenseEntitlementUnassignedWebhookPayload = {
  data: LicenseEntitlementWritable;
  /**
   * Event name
   */
  name: "LICENSE_ENTITLEMENT_UNASSIGNED";
  /**
   * Event type
   */
  type: "com.kaiten.license.entitlement.v1.unassigned";
};

export type OnLicenseEntitlementUnassignedWebhookRequest = {
  body: OnLicenseEntitlementUnassignedWebhookPayload;
  key: "com.kaiten.license.entitlement.v1.unassigned";
  path?: never;
  query?: never;
};

export type OnLicenseEntitlementUpdatedWebhookPayload = {
  data: LicenseEntitlementWritable;
  /**
   * Event name
   */
  name: "LICENSE_ENTITLEMENT_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.license.entitlement.v1.updated";
};

export type OnLicenseEntitlementUpdatedWebhookRequest = {
  body: OnLicenseEntitlementUpdatedWebhookPayload;
  key: "com.kaiten.license.entitlement.v1.updated";
  path?: never;
  query?: never;
};

export type OnLicenseArchivedWebhookPayload = {
  data: LicenseWritable;
  /**
   * Event name
   */
  name: "LICENSE_ARCHIVED";
  /**
   * Event type
   */
  type: "com.kaiten.license.v1.archived";
};

export type OnLicenseArchivedWebhookRequest = {
  body: OnLicenseArchivedWebhookPayload;
  key: "com.kaiten.license.v1.archived";
  path?: never;
  query?: never;
};

export type OnLicenseCreatedWebhookPayload = {
  data: LicenseWritable;
  /**
   * Event name
   */
  name: "LICENSE_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.license.v1.created";
};

export type OnLicenseCreatedWebhookRequest = {
  body: OnLicenseCreatedWebhookPayload;
  key: "com.kaiten.license.v1.created";
  path?: never;
  query?: never;
};

export type OnLicenseDeletedWebhookPayload = {
  data: LicenseWritable;
  /**
   * Event name
   */
  name: "LICENSE_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.license.v1.deleted";
};

export type OnLicenseDeletedWebhookRequest = {
  body: OnLicenseDeletedWebhookPayload;
  key: "com.kaiten.license.v1.deleted";
  path?: never;
  query?: never;
};

export type OnLicensePublishedWebhookPayload = {
  data: LicenseWritable;
  /**
   * Event name
   */
  name: "LICENSE_PUBLISHED";
  /**
   * Event type
   */
  type: "com.kaiten.license.v1.published";
};

export type OnLicensePublishedWebhookRequest = {
  body: OnLicensePublishedWebhookPayload;
  key: "com.kaiten.license.v1.published";
  path?: never;
  query?: never;
};

export type OnLicenseUnarchivedWebhookPayload = {
  data: LicenseWritable;
  /**
   * Event name
   */
  name: "LICENSE_UNARCHIVED";
  /**
   * Event type
   */
  type: "com.kaiten.license.v1.unarchived";
};

export type OnLicenseUnarchivedWebhookRequest = {
  body: OnLicenseUnarchivedWebhookPayload;
  key: "com.kaiten.license.v1.unarchived";
  path?: never;
  query?: never;
};

export type OnLicenseUpdatedWebhookPayload = {
  data: LicenseWritable;
  /**
   * Event name
   */
  name: "LICENSE_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.license.v1.updated";
};

export type OnLicenseUpdatedWebhookRequest = {
  body: OnLicenseUpdatedWebhookPayload;
  key: "com.kaiten.license.v1.updated";
  path?: never;
  query?: never;
};

export type OnLicenseFamilyCreatedWebhookPayload = {
  data: LicenseFamilyViewWritable;
  /**
   * Event name
   */
  name: "LICENSE_FAMILY_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.license_family.v1.created";
};

export type OnLicenseFamilyCreatedWebhookRequest = {
  body: OnLicenseFamilyCreatedWebhookPayload;
  key: "com.kaiten.license_family.v1.created";
  path?: never;
  query?: never;
};

export type OnLicenseFamilyDeletedWebhookPayload = {
  data: LicenseFamilyViewWritable;
  /**
   * Event name
   */
  name: "LICENSE_FAMILY_DELETED";
  /**
   * Event type
   */
  type: "com.kaiten.license_family.v1.deleted";
};

export type OnLicenseFamilyDeletedWebhookRequest = {
  body: OnLicenseFamilyDeletedWebhookPayload;
  key: "com.kaiten.license_family.v1.deleted";
  path?: never;
  query?: never;
};

export type OnLicenseFamilyUpdatedWebhookPayload = {
  data: LicenseFamilyViewWritable;
  /**
   * Event name
   */
  name: "LICENSE_FAMILY_UPDATED";
  /**
   * Event type
   */
  type: "com.kaiten.license_family.v1.updated";
};

export type OnLicenseFamilyUpdatedWebhookRequest = {
  body: OnLicenseFamilyUpdatedWebhookPayload;
  key: "com.kaiten.license_family.v1.updated";
  path?: never;
  query?: never;
};

export type OnMetadataFieldArchivedWebhookPayload = {
  data: MetadataFieldWritable;
  /**
   * Event name
   */
  name: "METADATA_FIELD_ARCHIVED";
  /**
   * Event type
   */
  type: "com.kaiten.metadata_field.v1.archived";
};

export type OnMetadataFieldArchivedWebhookRequest = {
  body: OnMetadataFieldArchivedWebhookPayload;
  key: "com.kaiten.metadata_field.v1.archived";
  path?: never;
  query?: never;
};

export type OnMetadataFieldCreatedWebhookPayload = {
  data: MetadataFieldWritable;
  /**
   * Event name
   */
  name: "METADATA_FIELD_CREATED";
  /**
   * Event type
   */
  type: "com.kaiten.metadata_field.v1.created";
};

export type OnMetadataFieldCreatedWebhookRequest = {
  body: OnMetadataFieldCreatedWebhookPayload;
  key: "com.kaiten.metadata_field.v1.created";
  path?: never;
  query?: never;
};

export type OnMetadataFieldReorderedWebhookPayload = {
  data: MetadataFieldReordered;
  /**
   * Event name
   */
  name: "METADATA_FIELD_REORDERED";
  /**
   * Event type
   */
  type: "com.kaiten.metadata_field.v1.reordered";
};

export type OnMetadataFieldReorderedWebhookRequest = {
  body: OnMetadataFieldReorderedWebhookPayload;
  key: "com.kaiten.metadata_field.v1.reordered";
  path?: never;
  query?: never;
};

export type OnMetadataFieldUnarchivedWebhookPayload = {
  data: MetadataFieldWritable;
  /**
   * Event name
   */
  name: "METADATA_FIELD_UNARCHIVED";
  /**
   * Event type
   */
  type: "com.kaiten.metadata_field.v1.unarchived";
};

export type OnMetadataFieldUnarchivedWebhookRequest = {
  body: OnMetadataFieldUnarchivedWebhookPayload;
  key: "com.kaiten.metadata_field.v1.unarchived";
  path?: never;
  query?: never;
};